+++ source/Messaging.IntegrationTests/Application/Transactions/MoveIn/CreateEndOfSupplyNotificationTests.cs
@@ -22,6 +22,7 @@ using Messaging.Application.Transactions;
 using Messaging.Application.Transactions.MoveIn;
 using Messaging.Domain.OutgoingMessages;
 using Messaging.Domain.Transactions.MoveIn;
+using Messaging.IntegrationTests.Assertions;
 using Messaging.IntegrationTests.Fixtures;
 using Xunit;
